# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Empleados",
        "description": "Documentación de la API para gestionar empleados con autenticación JWT.",
        "contact": {
            "email": "soporte@empresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/empleados": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener la lista de empleados con opción de filtrar por documento, nombre o email",
                "operationId": "01e9d10f171ec0d88e4010f4876a0b65",
                "parameters": [
                    {
                        "name": "buscar",
                        "in": "query",
                        "description": "Filtrar empleados por documento de identidad, nombre o email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de empleados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "documento_identidad": {
                                                "type": "string",
                                                "example": "123456789"
                                            },
                                            "nombres": {
                                                "type": "string",
                                                "example": "Juan"
                                            },
                                            "apellidos": {
                                                "type": "string",
                                                "example": "Pérez"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "juan@example.com"
                                            },
                                            "cargo": {
                                                "type": "string",
                                                "example": "Desarrollador"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "example": "activo"
                                            },
                                            "fecha_nacimiento": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1990-05-15"
                                            },
                                            "fecha_ingreso": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2022-06-01"
                                            },
                                            "fecha_creacion": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-02-17T03:37:53.000000Z"
                                            },
                                            "fecha_actualizacion": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-02-17T03:37:53.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear un nuevo empleado",
                "operationId": "8dbd970d554ec6064a0527ac54baa93f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "documento_identidad",
                                    "nombres",
                                    "apellidos",
                                    "email",
                                    "cargo",
                                    "fecha_nacimiento",
                                    "fecha_ingreso",
                                    "estado"
                                ],
                                "properties": {
                                    "documento_identidad": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "cargo": {
                                        "type": "string",
                                        "example": "Desarrollador"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "fecha_ingreso": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-06-01"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "documento_identidad": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Ya existe un empleado con este documento de identidad."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El correo electrónico ya está en uso."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/empleados/{id}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener los datos de un empleado por su ID",
                "operationId": "7d4cc7b8b0f8f7ee805539c3711aed25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del empleado obtenidos correctamente"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar los datos de un empleado",
                "operationId": "032112c4832b57bdfbf99a313486352d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "documento_identidad": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "cargo": {
                                        "type": "string",
                                        "example": "Gerente de TI"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "fecha_ingreso": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-06-01"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "activo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado correctamente"
                    },
                    "404": {
                        "description": "Empleado no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Empleado no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "documento_identidad": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Este documento de identidad ya está registrado en otro empleado."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El correo electrónico ya está en uso."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar un empleado",
                "operationId": "cc4ed7cb13a8f08b869cb1b1775ec40f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado eliminado correctamente"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}